EDK461 EDK463 EDK468 EDK476 EDK479 EDK486 EDK487 EDK488 EDK490 EDK492 EDK493 EDK494 EDK495 EDK496 EDK497 EDK498 EDK500 EDK501 EDK502 EDK504 EDK505 EDK506 EDK507 EDK508 EDK510 EDK511 EDK512 EDK514 EDK515 EDK516 EDK517 EDK518 EDK520 EDK522 EDK524 EDK525 EDK526
git-svn-id: https://svn.code.sf.net/p/efidevkit/code/trunk/Edk@169 d4c01ab0-1994-4ed3-9bc5-3ada14f9202f

+++ Sample/Universal/Network/Tcp4/Dxe/Tcp4Timer.c
@@ -156,7 +156,7 @@ Returns:
 
 --*/
 {
-  int FlightSize;
+  UINT32  FlightSize;
 
   TCP4_DEBUG_WARN (("TcpRexmitTimeout: transmission "
     "timeout for TCB %x\n", Tcb));
@@ -167,7 +167,7 @@ Returns:
   // yet ACKed.
   //
   FlightSize        = TCP_SUB_SEQ (Tcb->SndNxt, Tcb->SndUna);
-  Tcb->Ssthresh     = NET_MAX (2 * Tcb->SndMss, FlightSize / 2);
+  Tcb->Ssthresh     = NET_MAX ((UINT32) (2 * Tcb->SndMss), FlightSize / 2);
 
   Tcb->CWnd         = Tcb->SndMss;
   Tcb->LossRecover  = Tcb->SndNxt;
